Mise en place d'une bonne partie de l'application.
Il reste le teste des performances
Corriger le problème de sacade lors d'un changement de page
Mise en place correct des fonctions audio (bouton etc..)

+++ SoundLink_Server/soundlink_dao/src/main/java/soundlink/dao/repositories/ArtisteRepository.java
@@ -10,7 +10,7 @@ import soundlink.model.entities.Artiste;
  * @author xrobert
  *
  */
-public interface ArtisteRepository extends JpaRepository<Artiste, Long> {
+public interface ArtisteRepository extends JpaRepository<Artiste, Integer> {
 
     /**
      * Allow you to find an artiste by is name
